+++ FirebaseAuth/Sources/Swift/Backend/RPC/EmailLinkSignInRequest.swift
@@ -32,7 +32,6 @@ private let kPostBodyKey = "postBody"
 /// The key for the tenant id value in the request.
 private let kTenantIDKey = "tenantId"
 
-@available(iOS 13, tvOS 13, macOS 10.15, macCatalyst 13, watchOS 7, *)
 class EmailLinkSignInRequest: IdentityToolkitRequest, AuthRPCRequest {
   typealias Response = EmailLinkSignInResponse