+++ FirebaseAuth/Sources/Swift/Backend/RPC/GetRecaptchaConfigRequest.swift
@@ -43,7 +43,6 @@ private let kVersionKey = "version"
 /// The key for the tenant id value in the request.
 private let kTenantIDKey = "tenantId"
 
-@available(iOS 13, tvOS 13, macOS 10.15, macCatalyst 13, watchOS 7, *)
 class GetRecaptchaConfigRequest: IdentityToolkitRequest, AuthRPCRequest {
   typealias Response = GetRecaptchaConfigResponse