+++ FirebaseAuth/Tests/Unit/GoogleAuthProviderTests.swift
@@ -20,7 +20,6 @@ import XCTest
 /** @class FIRGoogleAuthProviderTests
     @brief Tests for @c FIRGoogleAuthProvider
  */
-@available(iOS 13, tvOS 13, macOS 10.15, macCatalyst 13, watchOS 7, *)
 class GoogleAuthProviderTests: XCTestCase {
   let kAccessToken = "Token"
   let kIDToken = "idToken"